manim-worker/requirements.txt
manim
flask
python-dotenv
vercel-blob
pycairo
